[Fix #294]
Found issue where attempting to call `._init()` on the native module
threw a red screen of death in React Native v0.54.

In the docs for RN v.54, the tutorial says nothing about the RCT
prefix. Removing this, inline with the current documentation, seems to
have fixed the problem.

There seemed to be issues with the gradle file, preventing builds on
Android. I had to bump the compiler SDK versions for it to build so I
could test. The `Build.VERSION_CODES.O` constant is apparently only
available after API level 26. Changing the compiler SDK version,
whilst keeping the target the same fixed the build.